actions/requirements.txt
fastapi
pydantic
psycopg2
asyncpg
transformers
torch
python-dotenv
numpy==1.26.4

langchain
langchain_community
langchain_core
langchain_postgres 
langchain_cohere
pandas